backend_django/requirements.txt
Django
django-cors-headers
psycopg2
djangorestframework